viscode-api-server/requirements.txt
flask
pymongo
elasticsearch
psycopg2-binary
requests
python-dotenv